[BigQuery] Expand _get_all_addresses_between_start_and_end_collections to accept source tables (Recidiviz/recidiviz-data#33865)
## Description of the change

Makes a small update to the logic we use to select the full set of
addresses to load via `auto` mode in `load_views_to_sandbox`. Previously
we could only denote *view* addresses as addresses to start exploration
from, but now we can also pass in a list of source table addresses that
we should start exploration from.

This PR also moves the functionality into the `BigQueryDagWalker` and
adds unittests.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#33733

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a420d1bd8cd3a7029cc5fc4a071e00e6abaebbda