[BigQuery] Eliminate all legacy EventQueryBuilder and SpanQueryBuilder logic (Recidiviz/recidiviz-data#33848)
## Description of the change

Follow-up to Recidiviz/recidiviz-data#33821 - now that the new `all_person_events`, etc views
are fully hydrated, we can use these in place of the legacy
`person_events`, etc views.

This PR starts by just replacing the query text for the legacy views
with a `SELECT *` query on the new view that replaces it. We need to
keep these legacy views around for a week while the new query logic
ships to prod. Then we'll be able to delete references from Looker all
at once.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#32921

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c6be3bfbb01c7e67f623f0393326d95894d6aeac