[US_AZ][Workflows] Additional Criteria and AZ Query Fragments file (Recidiviz/recidiviz-data#33868)
## Description of the change

`Completed functional literacy`: `3565 people are no longer eligible`,
this is expected because we are using a more accurate representation of
the literacy data whereas the previous dataset we were using had quite a
few issues based on conversations with Trusted Testers, so we expect to
see many less eligible individuals but more accuracy for those that are.
I will do validations on this criteria in particular to ensure this
large shift is something we expect to see.

`No Drug Offense Conviction`: This encompasses all four of the drug
conviction criteria, `107 residents are no longer eligible` as a result.

`No Domestic Violence Conviction`: `21 residents no longer eligible`

`No Sexual Exploitation of Children`: No change, this is likely because
we already check for sexual-related crimes and such anyone with a
conviction for sexual exploitation has a conviction for the more general
category of Sexual Crimes.

Overall, `3691` individuals are no longer eligible.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33386
Closes Recidiviz/recidiviz-data#33387
Closes Recidiviz/recidiviz-data#33388
Closes Recidiviz/recidiviz-data#33390
Closes Recidiviz/recidiviz-data#33392

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4373a2ee5715c1d52ec548abf78dff61931bdfe8