[Justice Counts] Move agency_settings from Agency to Source in DB schema (Recidiviz/recidiviz-data#33928)
## Description of the change

This PR moves `agency_settings` from the Agency to the Source table in
the database. Agency is an abstraction of Source, so the tables are the
same. Because of this, when I ran pipenv run migrate-jc, the migration
file came up empty.

Why is this necessary? I’ll be manually creating sources to backfill
reporting_agency for new agency dashboards. These sources won't be
agencies; they’ll be vendors or CSG. To support this, I need to add a
"homepage URL" setting for these sources, which requires moving
agency_settings to Source.

## Related issues
Recidiviz/recidiviz-data#33842

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d4a90a3302bce602fb69ad7c881a2c8ba8e0d36f