[US_MI] Add columns to ADH_SHPERSON raw data config (Recidiviz/recidiviz-data#33918)
## Description of the change

Michigan added two columns to the `ADH_SHPERSON` raw data table on
10/3/24. Since we infer columns from raw data configs in MI, it is
unclear what the columns were that they added. This PR adds two new
columns `Unknown1` and `Unknown2` to the raw data config so that this
file can successfully import.

I tested import in a sandbox to view the file contents
(`elise_us_mi_raw_data.ADH_SHPERSON`). The two new columns are always
empty. Based on a search of MI raw data uses, it seems like the only
fields we use from this file are `RecID` and `IsDeleted`, both of which
appear unimpacted by this change. Those fields are used in ingest while
hydrating assessments, so if anything happens to assessments once this
newly-structured file is imported, we should investigate this table.

Once @brendali121 is back or I find out who our current data contact is
in MI, we should ask what these added columns were and rename the
columns in this config to reflect the actual structure of the file. I
made a ticket to do this at
[Recidiviz/recidiviz-data#33917](https://github.com/Recidiviz/recidiviz-data/issues/33917).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33915

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8ea78fdc85496b198aede0c9556732a614bbb594