[US_IX/ND] Change all references to transfer_to_reentry_enter to transfer_to_minimum_facility (Recidiviz/recidiviz-data#33351)
## Description of the change

Our name for this completion event has caused confusions at many levels,
so we thought it was a good time to change it's name to make it clearer.
By changing it to `transfer_to_minimum_facility` it both reflects what
it's capturing for both ID (transfer to CRC without work-release) and ND
(transfer to a minimum security facility).
## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e923321ee57266755ec2a34f649bcdb169cff239