[raw data import dag] change resource locking failure case, update dag sequencing (Recidiviz/recidiviz-data#33852)
## Description of the change

in this PR, i made a few updates to the sequencing and behavior of raw
data import DAG;
- if a failure occurs during resource locks acquisition, we skip all
downstream tasks
- if we fail to acquire to resource locks because they are already held,
we skip all downstream tasks but DONT fail -- in cases like colorado
where we want the infra to be always paused or want to prevent import
runs on purpose by holding onto a manual lock that overlaps with an
import (like raw data pruning), we dont want to fail loudly. if this is
unintended (we held on to a lock by accident), stale raw data validation
should be a backstop.

i also finally got rid of manually raising `AirflowSkipExceptions` which
i am happy about -- i feel they make it quite difficult to introspect
about things by just reading the DAG.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33849

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ba43785cae82497c877b0163221857c301953b51