[raw data import dag] ensure we are not re-importing processed files, consolidate blocking (Recidiviz/recidiviz-data#33893)
## Description of the change

first a few (i think three) PRs to update to operations registration
logic to make it handle more edge cases. this PR handles:
- a common case that i previously missed: if a file has already been
imported (i.e. has `file_processed_time`) we shouldnt re-import it
(folks should have to deprecate the file). technically we wouldnt be
duplicating data if we didnt do this since in the load step we delete
data w/ the same `file_id`
- validate that all of the chunks exist for chunked files that have
already been grouped in the past and have failed

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#33879

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 66c98fb863d2ccb09e5fada8cc66af2dc46b9b34