[Workflows] Preserve key order in response objects (Recidiviz/recidiviz-data#33957)
## Description of the change

Some opportunity config fields depend on dict key order, but by default
Flask sorts keys in response objects to improve cacheability. I have a
more permanent fix scoped in Recidiviz/recidiviz-data#33956, but there are too many moving parts
to shoot for today's RC. For this next week we can temporarily disable
the sorting behavior. It's apparently only possible to do this for the
entire server at once, but it's unlikely that we're relying on the
sorting behavior elsewhere.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b95d821709a5f3e8d8aaa6cbae3e3cde00497bc0