[admin panel] fix dag links (Recidiviz/recidiviz-data#33977)
## Description of the change

i was looking at the new detail pages before flashing (see
[example](https://admin-panel-staging.recidiviz.org/admin/ingest_operations/ingest_pipeline_summary/US_OZ/instance/SECONDARY/egt_person))
and noticed that the links didnt actually link to the right dag run.

i must have not tested it well enough? seems like go links will
automatically url-decode placeholders, i opened an issue in
trotto/go-linksRecidiviz/recidiviz-data#245 with more info.

for now, just manually have the airflow uris in code (although now that
i'm typing this out is it bad to have this in code since it'll be
mirrored in the public repo???????? ik it's behind IAM but still)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 931607127512af4ed04fd3f4df76f003ea112d30