[raw data import] add max update_datetime validation (Recidiviz/recidiviz-data#33934)
## Description of the change

adds an additional filter during the operations registration that
validates that files we are trying to import have an `update_datetime`
that is greater than the current max processed `update_datetime`. this
is supposed to block the case where we upload old data out of order

there are two additional types of "we are deprecating but not
invalidating files" that i did not write validations for relating to
unprocessed files. we can still go ahead with the import safely since
none of the data is in BQ (they are unprocessed), but the deprecated
files that are not present will be blocked if we do try to import them
at a later time --
- _case 1_: the min unprocessed `update_datetime` from the operations
database is less than the min `update_datetime` in the set of files to
import for the same file tag
- _how this would happen_: we failed to import a file and removed it
from the ingest bucket (i.e. has an invalid byte), but forgot to
invalidate it in the operations database.
- _if we were to block on this in this step of the DAG_: we would fail
the imports of the all files in favor of waiting until this file is
present
- _how we'd validate it_: add a `min_unprocessed_update_datetime` field
to the `MAX_PROCESSED_UPDATE_DATETIME_FOR_FILE_TAG` query
- _impact of not validating_: we'd mainly just have a stray
non-invalidated file that wasn't processed. only time we'd really notice
is during flashing, but maybe we should set up an actual validation
around this. importing would be blocked of the missing file if we did
try to reimport it via the max `update_datetime` validation included in
this PR.
- _case 2_: there's a file in the operations database that is not
present in the ingest bucket with an `update_datetime` that is greater
than the min unprocessed `update_datetime` in the set of files to import
for the same file tag. we cant import any after
- _how this would happen_: tbh not totally sure on this one, but stuff
happens idk raw data is the wild west
- _if we were to block on this in this step of the DAG_: we would fail
the imports of the all files after the missing file
- _how we'd validate it_: add a new query that compares the set of all
non-processed, non-invalidated files with the set of files we are trying
to import
- _impact of not validating_: we'd mainly just have a stray
non-invalidated file that wasn't processed. only time we'd really notice
is during flashing, but maybe we should set up an actual validation
around this. importing would be blocked of the missing file if we did
try to reimport it via the max `update_datetime` validation included in
this PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#33879

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0962424a2f3701e5c9f3b12b7d5a7e3db0c4c501