[raw data import] add `raise_registration_errors` task (Recidiviz/recidiviz-data#33970)
## Description of the change

instead of just logging errors, this PR collects & push them to xcom to
be raised in `raise_registration_errors`. this way, we can feed the
output of this xcom into a our new alerting system (whatever that may
be) to have these errors included in pagerduty/github

because not all of these files have been registered with the operations
db, i don't think we can add a new status to
`DirectIngestRawFileImportStatus` for them and so they don't fit into
our `RawDataImportError` model; we could change the
`direct_ingest_raw_file_import` table to have `file_id` fk be nullable
and also link `gcs_file_id` but i think that would be quite a lot and
also confusing.

ALSO -- `BaseOperators` don't have the `multiple_outputs` arg (coming in
airflow 2.9.0) but we can push something to xcom to a specific key and
then also have a return value which sort of acts like it. i personally
didn't find this approach unreadable or confusing, but i'm also a little
too deep into airflow to know if it is so pls lmk if just putting it
into a dictionary and then reading the key out would be better

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33879

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e2b05675304f1f92353a34eaf08ff2d7d6b61018